+++ ios/docs/html/search/typedefs_63.js
@@ -8,6 +8,7 @@ var searchData=
   ['cmstoreobjectuploadcallback',['CMStoreObjectUploadCallback',['../_c_m_store_callbacks_8h.html#a31bd90b82d404aa9586fc609fa21dd41',1,'CMStoreCallbacks.h']]],
   ['cmuserfetchcallback',['CMUserFetchCallback',['../_c_m_user_8h.html#abd3eece586565d4c1866f7a885b269e0',1,'CMUser.h']]],
   ['cmuseroperationcallback',['CMUserOperationCallback',['../_c_m_user_8h.html#a103162e6c8dfb13dea703dc3d31664a8',1,'CMUser.h']]],
+  ['cmwebservicedevicetokencallback',['CMWebServiceDeviceTokenCallback',['../_c_m_device_token_result_8h.html#ad28e348d87c8c88d5b7cd3fd3e3311c0',1,'CMDeviceTokenResult.h']]],
   ['cmwebservicefetchfailurecallback',['CMWebServiceFetchFailureCallback',['../_c_m_web_service_8h.html#a57f23dfaa54b80c1798acee35e45f386',1,'CMWebService.h']]],
   ['cmwebservicefilefetchsuccesscallback',['CMWebServiceFileFetchSuccessCallback',['../_c_m_web_service_8h.html#ae8267288b3502db7c66608bd3a43cc6a',1,'CMWebService.h']]],
   ['cmwebservicefileuploadsuccesscallback',['CMWebServiceFileUploadSuccessCallback',['../_c_m_web_service_8h.html#a5609890cd4b9e8590aaf19c69bf4dc3a',1,'CMWebService.h']]],